OpenDataEuskadiAPIs_V6/requirements.txt
flask
requests
folium
geopy
python-arango
